% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/agglomer.r
\name{agglomer}
\alias{agglomer}
\title{Partition the peak lists pooled by \code{\link{startprofiles}}.}
\usage{
agglomer(profileList, dmass = 2, ppm = TRUE, dret = 60)
}
\arguments{
\item{profileList}{A profile list.}

\item{dmass}{Numeric. m/z gap size}

\item{ppm}{Logical. \code{dmass} given in ppm?}

\item{dret}{Numeric. RT gap size; units equal to those of the input files}
}
\value{
Updated profile list
}
\description{
\code{agglomer} partitions all peaks from various files into unassociated regions.
}
\details{
enviMass workflow function
}
\seealso{
\code{startprofiles}, \code{partcluster}
}

