% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enviMass.r
\docType{package}
\name{enviMass}
\alias{enviMass}
\alias{enviMass-package}
\title{Trend detection in LC-MS data}
\description{
enviMass version 2 provides basic workflow functions and a webbrowser user interface for the proccessing of high resolution
LC-MS data, with emphasis on trend detection in time series of LC-MS files.
}
\details{
Incorporates functions from several other packages for peak picking, adduct grouping, isotope pattern calculation and grouping, etc.
Has only been tested for Thermo Orbitrap data, centroided and converted to .mzXML via Proteo Wizard`s msconvert tool.
The workflow user interface (UI) has only been tested with google chrome and firefox, which must be set as default browsers.


Exported functions:
\describe{
\item{\code{\link{webMass}}}{Start enviMass Workflow GUI.}
\item{\code{\link{recalib}}}{Function for m/z and RT recalibration.}
\item{\code{\link{PWfile}}}{Thermo .raw file conversion and centroidization with ProteoWizard.}
\item{\code{\link{profiletopeak}}}{Summarise profile information.}
\item{\code{\link{plotprofiles}}}{Interactive profile viewer.}
\item{\code{\link{plotglobal}}}{Plot all time series from a list of profiles.}
\item{\code{\link{plotaprofile}}}{Plot profiles.}
\item{\code{\link{plotagroup}}}{Plot time series of component profiles.}
\item{\code{\link{intensup}}}{Trend detection and blind subtraction for time profiles.}
\item{\code{\link{getID}}}{Find smallest available ID.}
\item{\code{\link{search_peak}}}{Find matches with peaklist.}
\item{\code{\link{screening}}}{Screens a sample and blank/blind peaklist for compound matches.}
\item{\code{\link{startprofiles}}}{Initializes a profile list object.}
\item{\code{\link{agglomer}}}{Partition the pooled peak lists generated by \code{\link{startprofiles}}.}
\item{\code{\link{partcluster}}}{Extract profiles from peak set partitions generated by \code{\link{agglomer}}.}
\item{\code{\link{mass_dens}}}{Kernel-density based mass estimation with confidence intervals.}
\item{\code{\link{mass_int}}}{Plot mass against intensity.}
}
}

