% Generated by roxygen2 (4.0.2): do not edit by hand
\name{export_profile_list}
\alias{export_profile_list}
\title{Copy profile details into matrix}
\usage{
export_profile_list(profileList,profpeaks2,progbar = FALSE)
}
\arguments{
\item{profileList}{A profile list.}
\item{profpeaks2}{Character dataframe with filtered profile metadata.}
\item{progbar}{Logical. Should a progress bar be shown? Only for Windows.}
}
\value{
Matrix of numerics
}
\description{
Write mean mass, mean RT, ID and peak intensities of profiles into a single matrix
}
\details{
enviMass workflow export function
}

