% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getID.r
\name{getID}
\alias{getID}
\title{Find smallest available (file, ...) ID}
\usage{
getID(IDs)
}
\arguments{
\item{IDs}{Vector of integers}
}
\description{
Given a list of numerical IDs, \code{IDs} returns the smallest integer not part of that list.
}
\details{
enviMass workflow function
}

