% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intensup.r
\name{intensup}
\alias{intensup}
\title{Trend detection and blind subtraction for time profiles.}
\usage{
intensup(profileList, from = FALSE, to = FALSE, progbar = FALSE,
  blindsub = TRUE, blindfold = 100, lags = c(5, 14), threshold = 3,
  notrend = FALSE)
}
\arguments{
\item{profileList}{A profile list.}

\item{from}{Logical or integer of index.}

\item{to}{Logical or integer of index.}

\item{progbar}{Logical. Should a progress bar be shown? Only for Windows.}

\item{blindsub}{Logical. Run blind subtraction?}

\item{blindfold}{Numerical. Blind definition; above blind, if \code{blindfold} larger in intensity.}

\item{lags}{Vector of numericals.}

\item{threshold}{Numerical. A trend is reported if its intensity is \code{threshold} above the mean intensity plus the intensity deviation of other trends.}

\item{notrend}{Logical. Report global trend intensity as maximum intensity after blind subtraction.}
}
\value{
Updated \code{profileList[[7]]}.
}
\description{
\code{intensup} runs a trend detection and blind subtraction for the list of time profiles.
}
\details{
enviMass workflow function for trend detection and blind subtraction.
}

