% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{IS}
\alias{IS}
\title{Exemplary data set of internal standard (IS) compounds.}
\format{A data frame with 19 rows and 16 variables}
\usage{
data(IS)
}
\description{
\itemize{
  \item ID. Unique compound IDs - character.
  \item Name. Commpound name - character.
  \item Formula. Compound molecular formula - character.
  \item RT. Retention time - numeric or character.
  \item RT_tolerance. Character of RT Tolerance or FALSE if default used.
  \item main_adduct. Name of restrictive main adduct or FALSE if default used.
  \item ion_mode. "positive" or "negative".
  \item Use_for_recalibration. Logical, character.
  \item Use_for_screening. Logical, character.
  \item restrict_adduct. Logical, character.
  \item Remark. Character.
  \item tag1. Character.
  \item tag2. Character.
  \item tag3. Character.
  \item from. POSIXct date, time - character.
  \item to. POSIXct date, time - character.
}
}
\keyword{datasets}

