% Generated by roxygen2 (4.0.2): do not edit by hand
\name{newproject}
\alias{newproject}
\title{Start new enviMass workflow project}
\usage{
newproject(pro_name, pro_dir, IS, targets)
}
\arguments{
\item{pro_name}{Character string of project name.}

\item{pro_dir}{Character string. Path to a valid folder that will contains the project folder.}

\item{IS}{Dataframe containing internal standard (IS) compounds}

\item{targets}{Dataframe containing target and suspect compounds}
}
\description{
\code{newproject} initializes an enviMass project
}
\details{
enviMass workflow function. Creates a project folder, various subfolders, a logfile, dummy outputs, etc
}

