% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_profiles_intensity_histograms}
\alias{plot_profiles_intensity_histograms}
\title{Plot intensities histogram of a profile}
\usage{
plot_profiles_intensity_histograms(mean_intensities, max_intensities, past_incidents,
current_incidents)
}
\arguments{
\item{mean_intensities}{Numeric vector of mean intensities in profiles}
\item{max_intensities}{Numeric vector of max intensities in profiles}
\item{past_incidents}{Numeric vector of global trands in profiles}
\item{current_incidents}{Numeric vector of current trends in profiles}
}
\description{
\code{plot_profile_intensities_histogram} plots the intensities histogram of given profiles
}
\details{
enviMass workflow plot function
}

