% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotagroup.r
\name{plotagroup}
\alias{plotagroup}
\title{Plot time series of component profiles}
\usage{
plotagroup(components, ID, profileList, logint = FALSE, blindsub = TRUE,
  blindfold = 100, lags = c(5, 14), threshold = 3, add = FALSE,
  textit = FALSE)
}
\arguments{
\item{components}{List of components}

\item{ID}{Component ID}

\item{profileList}{A profile list.}

\item{logint}{Logical. Plot parameter.}

\item{blindsub}{Logical. Plot parameter.}

\item{blindfold}{Numeric. Plot parameter.}

\item{lags}{Numeric vector. Plot parameter.}

\item{threshold}{Numeric. Plot parameter.}

\item{add}{Logical. Plot parameter.}

\item{textit}{Logical. Plot parameter.}
}
\description{
\code{plotagroup} extracts all profiles part of a component with a certain ID and plots them.
}
\details{
enviMass workflow plot function
}

