% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotaprofile}
\alias{plotaprofile}
\title{Plot profiles}
\usage{
plotaprofile(profileList, profileID, logint = FALSE, blindsub = TRUE,
  blindfold = 100, lags = c(5, 14), threshold = 3, add = FALSE,
  textit = TRUE, simple = FALSE, colorit = FALSE, use_lwd = FALSE)
}
\arguments{
\item{profileList}{A profile list.}

\item{profileID}{ID of profile to be plotted}

\item{logint}{Logical. Use log10 intensities?}

\item{blindsub}{Logical. Plot blind subtraction?}

\item{blindfold}{Numerical.}

\item{lags}{Vector of numericals.}

\item{threshold}{Numerical}

\item{add}{Logical. Plot parameter. Add to current plot?}

\item{textit}{Logical. Plot parameter.}

\item{simple}{Logical. Plot parameter.}

\item{colorit}{Logical. Plot parameter.}

\item{use_lwd}{Logical. Plot parameter.}
}
\value{
A dataset.
}
\description{
Plot profiles
}
\details{
enviMass workflow function for profile plotting
}
\seealso{
\code{startprofiles}, \code{agglomer}, \code{partcluster}, \code{intensup}
}

