% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotglobale.r
\name{plotglobal}
\alias{plotglobal}
\title{Plot all time series from a list of profiles.}
\usage{
plotglobal(profileList, n_global = 10, n_present = 10,
  use_present = FALSE, highlight = FALSE)
}
\arguments{
\item{profileList}{A profile list.}

\item{n_global}{Numeric. How many global most intense profiles to include?}

\item{n_present}{Numeric. How many present most intense profiles to include?}

\item{use_present}{Logical.}

\item{highlight}{Logical.}
}
\description{
\code{plotglobal} plots all or a subset of most intense time series contained in \code{plofileList}
}
\details{
enviMass workflow function
}

