% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotprofiles.r
\name{plotprofiles}
\alias{plotprofiles}
\title{Interactive profile viewer}
\usage{
plotprofiles(profileList, RTlimit = FALSE, mzlimit = FALSE,
  timelimit = FALSE, ppmbar = 5, intlog = FALSE)
}
\arguments{
\item{profileList}{A profile list}

\item{RTlimit}{Restrict initial RT range; numerical vector of size two.}

\item{mzlimit}{Restrict initial m/z range; numerical vector of size two.}

\item{timelimit}{Restricts time limit.}

\item{ppmbar}{Size of ppm bar of m/z-axis; only plotted when zoomed in.}

\item{intlog}{Should intensities be log10-transformed?}
}
\description{
\code{plotprofiles} provides interactive access to all profiles and the underlying picked peaks.
}
\details{
Interactive profile viewer; not part of workflow.
}

