% Generated by roxygen2 (4.0.2): do not edit by hand
\name{profiledist}
\alias{profiledist}
\title{Plot trend distributions}
\usage{
profiledist(profileList,ret)
}
\arguments{
\item{profileList}{profileList}
\item{ret}{Logical. Should outlier ranking be returned?}
}
\value{
The boxplot in grey shows the intensity distributions of all trends of concern, listing the IDs, mean masses (m/z) and mean retention time (RT)
of the profiles with the most intense trends on the right. Colored points are used to elucidate the current
trend intensities from the latest input file. Red dots signify profiles with intensities in the outlier range
of the global (past and latest) trend intensities; green dots symbolize those below.
}
\description{
\code{profiledist} boxplots current and global trend distributions
}
\details{
enviMass workflow function
}

