% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/profiletopeak2.r
\name{profiletopeak2}
\alias{profiletopeak2}
\title{Summarise profile information}
\usage{
profiletopeak2(profileList, logfile, progbar, get_what = "peaks",
  at_ID = FALSE)
}
\arguments{
\item{profileList}{A profile list (profiled, not just loaded or partitioned).}

\item{logfile}{enviMass project logfile.}

\item{progbar}{Logical. Should a progress bar be shown? Only for Windows.}

\item{get_what}{To be completed ...}

\item{at_ID}{To be completed ...}
}
\value{
Matrix or integer
}
\description{
\code{profiletopeak2}
}
\details{
enviMass workflow function
}

