% Generated by roxygen2 (4.0.2): do not edit by hand
\name{search_peak}
\alias{search_peak}
\title{Match with peaklist}
\usage{
search_peak(peaklist, mz, dmz = 5, ppm = TRUE, RT, dRT)
}
\arguments{
\item{peaklist}{Matrix or data.frame of peaks. 3 columns: m/z, intensity, RT.}

\item{mz}{Numeric vector. Mass(es) to be matched.}

\item{dmz}{Numeric. +/- m/z tolerance (precision)}

\item{ppm}{Logical. \code{dmass} given in ppm?}

\item{RT}{Numeric vector. Retention times to be matched; units equal to those of the input files}

\item{dRT}{Numeric. RT tolerance window; units equal to those of the input files}
}
\value{
Vector of characters. For each code{mz,RT} input, the vector entry specifies the match indices in \code{peaklist}.
}
\description{
\code{search_peak} searchs for a peak with a certain mass and RT in a peaklist.
}
\details{
enviMass workflow function
}
\seealso{
\code{screening}
}

