% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/startprofiles.r
\name{startprofiles}
\alias{startprofiles}
\title{Initialize profile list object}
\usage{
startprofiles(logfile, frac = FALSE, sets = FALSE, progbar = FALSE,
  ion_mode = "positive", until = FALSE)
}
\arguments{
\item{logfile}{logfile object of an enviMass project.}

\item{frac}{Numerical, \code{0<frac=<1}. Fraction of files to use; oldest are ommited.}

\item{sets}{Integer. Number of latest files to include.}

\item{progbar}{Logical. Should a progress bar be shown? Only for Windows.}

\item{ion_mode}{Character string, either "positive" or "negative".}

\item{until}{Integer, ID of file. All peaks of files up to the date of this file will be included.}
}
\value{
profile list
}
\description{
\code{startprofiles} initializes a list object containing all peaks from the files in an enviMass project, associated metadata and placeholder.
}
\details{
enviMass workflow function
}
\seealso{
\code{agglomer}, \code{partcluster}
}

