% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{targets}
\alias{targets}
\title{Exemplary data set of target compounds.}
\format{A data frame with 68 rows and 19 variables}
\usage{
data(targets)
}
\description{
\itemize{
  \item ID. Unique compound IDs - character.
  \item Name. Commpound name - character.
  \item Formula. Compound molecular formula - character.
  \item RT. Retention time - numeric or character.
  \item RT_tolerance. Character of RT Tolerance or FALSE if default used.
  \item ID_internal_standard. ID of internal standard - first column of IS data.frame.
  \item main_adduct. Name of restrictive main adduct or FALSE if default used.
  \item ion_mode. "positive" or "negative".
  \item Use_for_recalibration. Logical, character.
  \item Use_for_screening. Logical, character.
  \item restrict_adduct. Logical, character.
  \item Remark. Character.
  \item tag1. Character.
  \item tag2. Character.
  \item tag3. Character.
  \item from. POSIXct date, time - character.
  \item to. POSIXct date, time - character.
  \item intercept. Quantization parameter - character.
  \item slope. Quantization parameter - character.
}
}
\keyword{datasets}

