% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkproject.r
\name{checkproject}
\alias{checkproject}
\title{Check enviMass project}
\usage{
checkproject(logfile, isotopes, adducts)
}
\arguments{
\item{logfile}{enviMass project logfile}

\item{isotopes}{Isotope list}

\item{adducts}{Adducts list from package enviPat}
}
\value{
Character string with either specific error message or a "Project consistent" message.
}
\description{
\code{} runs several cconsistency checks on a enviMass project
}
\details{
enviMass workflow function; run before further calculations are started in the workflow.
}

