% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convDate}
\alias{convDate}
\title{Convert dates}
\usage{
convDate(measurements)
}
\arguments{
\item{measurements}{enviMass file list}
}
\description{
\code{convDate} streamlines date inputs for interlan workflow usage
}
\details{
enviMass workflow function
}

