% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mass_dens.r
\name{mass_dens}
\alias{mass_dens}
\title{Mass density of a profile}
\usage{
mass_dens(profileList, profileID, bootstrap = FALSE, boot_size = 200,
  use_weights = FALSE)
}
\arguments{
\item{profileList}{A profile list.}

\item{profileID}{ID of profile to be plotted}

\item{bootstrap}{Add bootstrap confidene intervals?}

\item{boot_size}{Size of bootstrap sample}

\item{use_weights}{Weight by intensity?}
}
\description{
Plot mass density of a profile
}
\details{
enviMass workflow function for profile plotting
}

