% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_profiles_intensity_histograms.r
\name{plot_profiles_intensity_histograms}
\alias{plot_profiles_intensity_histograms}
\title{Plot intensities histogram of a profile}
\usage{
plot_profiles_intensity_histograms(mean_intensities, max_intensities,
  past_incidents, current_incidents)
}
\arguments{
\item{mean_intensities}{Numeric vector of mean intensities in profiles}

\item{max_intensities}{Numeric vector of max intensities in profiles}

\item{past_incidentss}{Numeric vector of global trands in profiles}

\item{current_incidentss}{Numeric vector of current trends in profiles}
}
\description{
\code{plot_profile_intensities_histogram} plots the intensities histogram of given profiles
}
\details{
enviMass workflow plot function
}

