% Generated by roxygen2 (4.0.2): do not edit by hand
\name{profiletopeak}
\alias{profiletopeak}
\title{Summarise profile information}
\usage{
profiletopeak(profileList, progbar)
}
\arguments{
\item{profileList}{A profile list.}

\item{progbar}{Logical. Should a progress bar be shown? Only for Windows.}
}
\value{
Matrix. See \code{colnames} of return for more information.
}
\description{
\code{profiletopeak} converts information from the profile list into a matrix, partly resembling
a matrix of peaks (often called peaklist).
}
\details{
enviMass workflow function
}

