% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PWfile.r
\name{PWfile}
\alias{PWfile}
\title{Thermo .raw file conversion and centroidization with ProteoWizard (PW)}
\usage{
PWfile(infile, folderout, msconvert_path, notintern = FALSE,
  use_format = "mzXML")
}
\arguments{
\item{infile}{Path to input file}

\item{folderout}{Path to output folder}

\item{msconvert_path}{Path to PW msconvert executable (including \\msconvert).}

\item{notintern}{Ignore}

\item{use_format}{Output format}
}
\description{
\code{PWfile} calls PW msconvert
}
\details{
enviMass workflow function. PW (not just msconvert) needs to be installed seperately, including the vendor library.
}

