% Generated by roxygen2 (4.0.2): do not edit by hand
\name{webMass}
\alias{webMass}
\title{enviMass Workflow GUI}
\usage{
webMass()
}
\description{
\code{webMass()} starts the browser user interface.
}
\details{
Requires a browser other than Internet Explorer. Has only been tested on Firefox and Google Chrome;
make sure to set the relevant browser as your default browser.
}

